backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-core==0.1.10
langchain-community==0.0.13
pinecone-client==3.1.0
python-multipart==0.0.6
httpx==0.26.0
aiohttp==3.9.1
chromadb==0.4.18
PyMuPDF==1.23.7
pdf2image==1.16.3
python-docx==1.0.1
Pillow==10.1.0 